requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
greenlet==2.0.2
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.6
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pywin32==306
pyzmq==25.0.2
six==1.16.0
SQLAlchemy==2.0.8
SQLAlchemy-Utils==0.40.0
stack-data==0.6.2
tblib==1.7.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
wcwidth==0.2.6
